back component wrote to TS
