+++ src/components/Back/Back.mdx
@@ -1,5 +1,5 @@
 import { Playground, Props } from 'docz'
-import Back from '@components/Back'
+import { Back } from '@components/Back'
 
 # Back
